+++ pkg-py/examples/01-hello-app.py
@@ -1,6 +1,5 @@
-from seaborn import load_dataset
 from querychat import QueryChat
+from querychat.data import titanic
 
-titanic = load_dataset("titanic")
-qc = QueryChat(titanic, "titanic")
+qc = QueryChat(titanic(), "titanic")
 app = qc.app()