+++ pkg-py/examples/03-sidebar-express-app.py
@@ -1,11 +1,9 @@
-from seaborn import load_dataset
 from shiny.express import render, ui
 from querychat.express import QueryChat
-
-titanic = load_dataset("titanic")
+from querychat.data import titanic
 
 # 1. Provide data source to QueryChat
-qc = QueryChat(titanic, "titanic")
+qc = QueryChat(titanic(), "titanic")
 
 # 2. Add sidebar chat control
 qc.sidebar()